Plot the missing data patterns for the optimal PM design.

plotPM plots the optimal PM design's missing patterns as a heatmap.

plotPM(object, colbr = "PRGn", col = c("antiquewhite1", "firebrick"),
  row.names = T, labels = T, fontsize_col = 20, fontsize_row = 14,
  fontsize = 14, angle_col = 45, legend = T, main = "", ...)

Arguments

object
The simPM object.

colbr
Specify the colors for different waves. Default is "PRGn".

col
Specify the colors for complete vs. missing data. Default is c("antiquewhite1","firebrick").

labels
logical scalar, indicating whether the label for waves is needed. Default is TRUE.

fontsize_col
Specify the font size for the column labels. Default is 20.

fontsize_row
Specify the font size for the row labels. Default is 14.

fontsize
Specify the font size for the legend. Default is 14.

angle_col
Specify the angle of how the column labels are displayed. Default is 45.

legend
Logical scalar, indicating whether the legend is shown. Default is TRUE.

main
Specify the plot title.

...
Any additional arguments for pheatmap.

See also

Examples

if (FALSE) {
plotPM(wave.out)
plotPM(indicator.out)
plotPM(forward.out,
labels=F,
col=c("gray96","gray35"),
fontsize_row=26,
fontsize=18,
fontsize_col=26)
}